# api.yaml
openapi: 3.0.3
info:
  title: KidsPOS API
  description: 子供向けPOSシステムのREST API
  version: 1.0.0
  contact:
    name: KidsPOS Team
    url: https://github.com/KidsPOSProject/KidsPOS-Server
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.kidspos.example.com
    description: Production server
...
paths:
...
  /api/sales:
    get:
      tags:
        - Sales
      summary: 売上一覧取得
      description: 全売上データを取得します
      operationId: getAllSales
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SaleResponse'

    post:
      tags:
        - Sales
      summary: 売上登録
      description: 新しい売上を登録します
      operationId: createSale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSaleRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: 売上ID
                  amount:
                    type: integer
                    description: 合計金額
                  quantity:
                    type: integer
                    description: 商品数
                  deposit:
                    type: integer
                    description: 預かり金額
                  change:
                    type: integer
                    description: おつり
                  staffId:
                    type: integer
                    description: スタッフID
                  storeId:
                    type: integer
                    description: 店舗ID
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Processing error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/sales/{id}:
    get:
      tags:
        - Sales
      summary: 売上詳細取得
      description: 指定したIDの売上詳細を取得します
      operationId: getSaleById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaleResponse'
        '404':
          description: Sale not found
  /api/sales/validate-printer/{storeId}:
    get:
      tags:
        - Sales
      summary: プリンター設定確認
      description: 指定した店舗のプリンター設定を確認します
      operationId: validatePrinter
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  isValid:
                    type: boolean

  # Staff
...
components:
  schemas:
...
    SaleResponse:
      type: object
      properties:
        id:
          type: integer
          description: 売上ID
        storeId:
          type: integer
          description: 店舗ID
        storeName:
          type: string
          description: 店舗名
        staffId:
          type: string
          description: スタッフID
        staffName:
          type: string
          description: スタッフ名
        totalAmount:
          type: integer
          description: 合計金額（リバー）
        deposit:
          type: integer
          description: 預り金（リバー）
        change:
          type: integer
          description: お釣り（リバー）
        saleTime:
          type: string
          format: date-time
          description: 売上時刻
        items:
          type: array
          description: 売上商品リスト
          items:
            $ref: '#/components/schemas/SaleItemResponse'
      required:
        - id
        - storeId
        - storeName
        - staffId
        - staffName
        - totalAmount
        - deposit
        - change
        - saleTime
    SaleItemResponse:
      type: object
      properties:
        itemId:
          type: integer
          description: 商品ID
        itemName:
          type: string
          description: 商品名
        barcode:
          type: string
          description: バーコード
        quantity:
          type: integer
          description: 数量
        unitPrice:
          type: integer
          description: 単価（リバー）
        subtotal:
          type: integer
          description: 小計（リバー）
      required:
        - itemId
        - itemName
        - barcode
        - quantity
        - unitPrice
        - subtotal
    CreateSaleRequest:
      type: object
      properties:
        storeId:
          type: integer
          description: 店舗ID
        staffBarcode:
          type: string
          nullable: true
          description: スタッフバーコード（オプショナル）
        itemIds:
          type: string
          description: 商品IDのカンマ区切り文字列（例："1,2,3"）
          example: "1,2,3"
        deposit:
          type: integer
          minimum: 0
          description: 預かり金額
      required:
        - storeId
        - itemIds
        - deposit
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: エラーコード
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: エラーメッセージ
          example: "商品IDが空です。売上作成には少なくとも1つの商品が必要です"
        timestamp:
          type: string
          format: date-time
          description: エラー発生時刻
        path:
          type: string
          nullable: true
          description: リクエストパス
        details:
          type: object
          nullable: true
          description: 追加の詳細情報
          additionalProperties: true
      required:
        - code
        - message
        - timestamp
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

# セキュリティは現在未実装のため、コメントアウト
# security:
#   - bearerAuth: []